Separate Master actor in two states
Start the Master in "Waiting for a Worker to appear" and a "Send work to workers"
state. This will make it easier to reconnect to workers.
Wait for the Master to be connected to a worker before sending work for
the Master to distribute among the workers.
Add more islands, now that we no longer parallelize the island evaluation.
This caused hangs for some reason I'm not interested in finding out.